pyproject.toml
[tool.poetry]
name = "api-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9, <4"
fastapi = "^0.81.0"
uvicorn = "^0.18.3"
pytest = "^7.1.3"
torch = "^1.12.1"
opencv-python = "^4.6.0.66"
Pillow = "^9.2.0"
requests = "^2.28.1"
torchvision = "^0.13.1"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
